+++ website/src/pages/index.tsx
@@ -55,7 +55,6 @@ function Home() {
 
                 <UITeaser/>
 
-                <FeaturesTeaser/>
 
                 <DeveloperFeatures/>